Removed 1-shot that stretched GT in Async Delays
If Rick/Josh are to be believed the GT pulse lasts for 60ns.
I previously believed this value to be 20ns. 20ns is to short to
be caught by the ds1023 so a 1-shot was added to stretch the pulse out.
However if GT is indeed 60ns this 1-shot is no longer needed.