Started trying layout and deleted 10E116 in comps.
I began messing with layout and a bunch of files got created...
not big suprise.
But the actual change is there was a 10E116/10H116 in Tune/Fast comp
that changed a single ended signal to differential. But a few commits
ago I made it so the differential signal was no longer needed (or wanted)
so I now got rid of the 116s