Added packages to LEDs & inductors & changed caps
CSMD0603 caps didn't have packages. And while I could have
gotten them easily I instead decided that I would just use
0805 caps everywhere.

Additionally LEDS and Inductors didn't have a good package associated
with them. So I added a package that was lying around. But I didn't want
an SMD package for the LEDs so I had to change every led in the schematic
which were LED_1206s (which is SMD) to LED_l (which is not smd)