Added more space between ICs; Began tunecomp.
I added space so fingers could fit. Idk if I added enough but in general
im going with about 0.1-0.2 inches btwn chips. Also I began tunecomp.
Schematics required adding a pulldown to the makeshift ecl one-shot. And
changing the package on an F06 to SOIC rather than DIP. I still need to
add caps for the power pins on everything on that page.