Layout and set power pin for new comparators ICs.
I changed the power pins on each of the new components so that
VEE->VEE and VCC->GND as which is the ECL way (I think). I also
did the layout for the new comparator design. It'll likely change again
now that the power pins are right b/c a few components can be combined
now.