In Generic Utilities Ports used new input method
The method allows for the input signal to be optionally differential
or single ended. If single ended the inverting input gets pulled up to VBB
through a (currently) 50ohm resistor. If differential the pull up should
act as a pull up normally does for an ECL signal. I want to test this
method to figure out if it works and if so what size resistor can I use.
Rick said up to about 5k ohms should work. And obviously the higher the
resistance  the lower the current used (which is ideal)