In GT_Delays used GTRIG_TTL instead of GTRIG_ECL
Previously I had a ttl and ecl version coming into TUBII from the PORTS
block. The ttl version went to the MZ and the ecl to CAEN, GT_DELAYS &
ECAL_CNTRL. The GT_Delays immediatly translated it to TTL though. So I
figured why do the same thing twice and removed the TTL translation in
GT_DELAYS and had the ttl copy go to the block